PreTrain-Data_Reduction/requirements.txt
imagehash~=4.3.1
numpy~=1.22.3
scikit-learn~=1.1.1
pillow~=9.2.0
seaborn~=0.11.2
